+++ service-implementation/src/main/java/com/liaison/mailbox/service/dto/configuration/ProcessorLegacyDTO.java
@@ -26,6 +26,7 @@ import com.liaison.mailbox.dtdm.model.Processor;
 import com.liaison.mailbox.dtdm.model.ProcessorProperty;
 import com.liaison.mailbox.dtdm.model.ScheduleProfileProcessor;
 import com.liaison.mailbox.enums.EntityStatus;
+import com.liaison.mailbox.enums.ProcessorType;
 import com.liaison.mailbox.enums.Protocol;
 import com.liaison.mailbox.service.dto.configuration.request.RemoteProcessorPropertiesDTO;
 import com.liaison.mailbox.service.exception.MailBoxConfigurationServicesException;
@@ -126,6 +127,11 @@ public class ProcessorLegacyDTO extends ProcessorDTO {
 			GenericValidator validator = new GenericValidator();
             validator.validate(propertiesDTO);
             
+            if (ProcessorType.REMOTEUPLOADER.equals(processor.getProcessorType()) ||
+                    ProcessorType.REMOTEDOWNLOADER.equals(processor.getProcessorType())) {
+                MailBoxUtil.constructURLAndPort(propertiesDTO);
+            }
+            
 			if (null != propertiesDTO) {
 				String propertiesJSON = MailBoxUtil.marshalToJSON(this.getRemoteProcessorProperties());
 				processor.setProcsrProperties(propertiesJSON);
@@ -178,11 +184,11 @@ public class ProcessorLegacyDTO extends ProcessorDTO {
 
 				property = new ProcessorProperty();
 				propertyDTO.copyToEntity(property, false);
-				properties.add(property);
 				property.setProcessor(processor);
+				properties.add(property);
 			}
 			if (!properties.isEmpty()) {
-				processor.setDynamicProperties(properties);
+				processor.getDynamicProperties().addAll(properties);
 			}
 
 			// Set the protocol