+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@
 * GMB-708 - Wrong Status code returns in response when the user configured the bind users to the Mailbox ID's
 * GMB-718 - Set upper limit for grouping the files in Directory Sweeper
 * GMB-634 - Mailbox Uploaders should check the local payload and try to connect the SFTP server once the files are available to upload
+* GMB-646 - Validation is missing for processor properties in server side
 
 4.0.0
 -------------------------------